created a utility file to manage diff functions, reduced codes from app.js
